PR for reusing sklearn clustering methods
