+++ flair/__init__.py
@@ -23,8 +23,7 @@ from . import (  # noqa: E402 import after setting device
     models,
     nn,
     trainers,
-    visual,
-    clustering
+    visual
 )
 
 __version__ = "0.10"
