--- flair/clustering/__init__.py
@@ -1,5 +1,0 @@
-from .kmeans import KMeans
-from .birch import Birch
-from .em import ExpectationMaximization
-
-__all__ = ["KMeans", "Birch", "ExpectationMaximization"]
