--- flair/clustering/birch/__init__.py
@@ -1,9 +1,0 @@
-from .birch import Birch
-
-__all__ = ["Birch"]
-
-#TODO to be removed
-branching_factor_non_leaf: int = 3
-branching_factor_leaf: int = 2
-distance_max: float = 1000000000
-threshold: float = 0.01
