+++ flair/models/__init__.py
@@ -7,7 +7,7 @@ from .relation_extractor_model import RelationExtractor
 from .sequence_tagger_model import MultiTagger, SequenceTagger
 from .tars_model import FewshotClassifier, TARSClassifier, TARSTagger
 from .text_classification_model import TextClassifier
-from .clustering import to_clustering_dataset
+from .clustering import ClusteringModel
 
 __all__ = [
     "DependencyParser",
@@ -22,5 +22,5 @@ __all__ = [
     "TARSClassifier",
     "TARSTagger",
     "TextClassifier",
-    "to_clustering_dataset"
+    "ClusteringModel"
 ]
